add ConflictNoteState model type and move note helpers to Note.ts
