+++ packages/lib/BaseModel.ts
@@ -27,6 +27,7 @@ export enum ModelType {
 	SmartFilter = 15,
 	Command = 16,
 	NoteEmbedding = 17,
+	ConflictNoteState = 18,
 }
 
 export interface SearchOptions {
@@ -88,6 +89,7 @@ class BaseModel {
 		['TYPE_SMART_FILTER', ModelType.SmartFilter],
 		['TYPE_COMMAND', ModelType.Command],
 		['TYPE_NOTE_EMBEDDING', ModelType.NoteEmbedding],
+		['TYPE_CONFLICT_NOTE_STATE', ModelType.ConflictNoteState],
 	];
 
 	private static uuidGenerator: ()=> string = uuid.create;
@@ -109,6 +111,7 @@ class BaseModel {
 	public static TYPE_SMART_FILTER = ModelType.SmartFilter;
 	public static TYPE_COMMAND = ModelType.Command;
 	public static TYPE_NOTE_EMBEDDING = ModelType.NoteEmbedding;
+	public static TYPE_CONFLICT_NOTE_STATE = ModelType.ConflictNoteState;
 
 	// eslint-disable-next-line @typescript-eslint/no-unsafe-function-type -- Set by the app to redux dispatch; per-app action types diverge so the function is typed loosely here
 	public static dispatch: Function = function() {};
